Wed, May 24, 2023, 6:46 PM +03:00

+++ interpreter/interpreter.py
@@ -3,24 +3,24 @@ def main():
     x, y, z = expression.split(" ")
 
     if y == "+":
-        result = x + z
+        result = int(x + z)
         float(result)
         #round(result,1)
         print(result)
     elif y == "-":
-        result = x - z
+        result = int(x - z)
         float(result)
         #round(result,1)
         print(result)
     elif y == "/":
         z != 0
-        result = x / z
+        result = int(x / z)
         float(result)
         #round(result,1)
         print(result)
     elif y == "%":
         z != 0
-        result = x % z
+        result = int(x % z)
         float(result)
         #round(result,1)
         print(result)
